pyproject.toml
[tool.poetry]
name = "eventual-tortoise"
version = "0.5.0"
description = ""
authors = ["Ivan Dmitriesvkii <ivan.dmitrievsky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
tortoise-orm = "^0.17.2"
orjson = "^3.5.2"
eventual = "^0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.812"
pytest = "^5.2"
pytest-asyncio = "^0.14.0"
autoflake = "^1.4"
flake8 = "^3.9.1"
flake8-bugbear = "^21.4.3"
flake8-pie = "^0.8.1"
isort = "^5.8.0"
asynctest = "^0.13.0"
coverage = "^5.5"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
multi_line_output = 3
combine_as_imports = true

[tool.black]
target-version = ['py37']
